move token related code to dedicated module
