+++ src/lexer.rs
@@ -1,74 +1,6 @@
-use std::fmt::Debug;
-
 use crate::error::LexerError;
 use crate::scanner::{Scanner, ScannerState};
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TokenType {
-    LeftParen,         // (
-    RightParen,        // )
-    LeftBrace,         // {
-    RightBrace,        // }
-    LeftBracket,       // [
-    RightBracket,      // ]
-    LeftAngle,         // <
-    RightAngle,        // >
-    Comma,             // ,
-    Dot,               // .
-    Minus,             // -
-    Plus,              // +
-    Colon,             // :
-    Semicolon,         // ;
-    Slash,             // /
-    Star,              // *
-    Bang,              // !
-    Equal,             // =
-    Pipe,              // |
-    Percent,           // %
-    Hash,              // #
-    SingleQuote,       // '
-    DoubleQuote,       // "
-    DoubleLeftBrace,   // {{
-    DoubleRightBrace,  // }}
-    LeftBracePercent,  // {%
-    PercentRightBrace, // %}
-    LeftBraceHash,     // {#
-    HashRightBrace,    // #}
-    BangEqual,         // !=
-    DoubleEqual,       // ==
-    LeftAngleEqual,    // <=
-    RightAngleEqual,   // =>
-    Whitespace,        // special token to account for whitespace
-    Text,
-    Eof,
-}
-
-#[derive(Debug, Clone)]
-pub struct Token {
-    token_type: TokenType,
-    lexeme: String,
-    line: usize,
-}
-
-impl Token {
-    fn new(token_type: TokenType, lexeme: String, line: usize) -> Self {
-        Token {
-            token_type,
-            lexeme,
-            line,
-        }
-    }
-}
-
-pub trait Tokenizer: Scanner {
-    type Token: Debug;
-    type TokenType: Debug;
-    type Error: std::error::Error;
-
-    fn tokenize(&mut self) -> Result<Vec<Self::Token>, Self::Error>;
-    fn scan_token(&mut self) -> Result<(), Self::Error>;
-    fn add_token(&mut self, token_type: Self::TokenType);
-}
+use crate::token::{Token, TokenType, Tokenizer};
 
 pub struct Lexer {
     source: String,
